# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: CortexAI POS API
  description: |
    API REST do CortexAI POS — sistema de Ponto de Venda com inteligência artificial.
    Gerencia produtos, vendas, clientes, cupons, categorias, estoque e processamento de texto com IA.
  version: 6.0.0
  contact:
    name: CortexAI
servers:
  - url: http://localhost:3001
    description: Servidor de desenvolvimento
paths:
...
  /api/auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/LoginCredenciais'
                - $ref: '#/components/schemas/LoginPin'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  usuario:
                    type: object
                    properties:
                      id: { type: integer }
                      nome: { type: string }
                      login: { type: string }
                      role: { type: string }
                  token: { type: string }
        '401':
          description: Credenciais inválidas
  /api/auth/register:
    post:
      summary: Registrar novo usuário
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome, login, senha]
              properties:
                nome: { type: string }
                login: { type: string }
                senha: { type: string }
                pin: { type: string, maxLength: 4, minLength: 4 }
                role: { type: string, default: operador }
      responses:
        '200':
          description: Usuário registrado
        '400':
          description: Dados inválidos ou login já existe
  /api/auth/me:
    get:
      summary: Dados do usuário autenticado
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados do usuário

  # ====== PRODUTOS ======
...
  /api/categorias:
    get:
      summary: Listar todas categorias
      tags: [Categorias]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de categorias
  /api/categoria:
    post:
      summary: Criar nova categoria
      tags: [Categorias]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome: { type: string }
                descricao: { type: string }
                tipoProduto: { type: string }
                icone: { type: string }
                cor: { type: string }
                status: { type: string, default: ativa }
    put:
      summary: Atualizar categoria
      tags: [Categorias]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome: { type: string, description: 'Nome atual para identificar a categoria' }
                novoNome: { type: string }
                descricao: { type: string }
                status: { type: string }
    delete:
      summary: Deletar categoria
      tags: [Categorias]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome: { type: string }
      responses:
        '200':
          description: Categoria deletada

  # ====== VENDAS ======
...
  /api/clientes:
    get:
      summary: Listar clientes (com paginação)
      tags: [Clientes]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Busca por nome, CPF, email ou telefone
        - in: query
          name: pagina
          schema: { type: integer, default: 1 }
        - in: query
          name: limite
          schema: { type: integer, default: 50 }
        - in: query
          name: ativo
          schema: { type: string }
      responses:
        '200':
          description: Lista de clientes
  /api/cliente:
    post:
      summary: Criar novo cliente
      tags: [Clientes]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarCliente'
      responses:
        '200':
          description: Cliente criado
    put:
      summary: Atualizar cliente
      tags: [Clientes]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarCliente'
      responses:
        '200':
          description: Cliente atualizado
    delete:
      summary: Deletar cliente
      tags: [Clientes]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: integer }
      responses:
        '200':
          description: Cliente deletado

  # ====== CUPONS ======
  /api/cupons:
    get:
      summary: Listar todos cupons
      tags: [Cupons]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de cupons
  /api/cupom:
    post:
      summary: Criar novo cupom
      tags: [Cupons]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [codigo, tipo, valor, validoAte]
              properties:
                codigo: { type: string, example: DESCONTO10 }
                tipo: { type: string, example: percentual }
                valor: { type: number, example: 10 }
                valorMinimo: { type: number }
                validoAte: { type: string, format: date-time }
                usosMax: { type: integer, default: 1 }
      responses:
        '200':
          description: Cupom criado
    put:
      summary: Atualizar cupom
      tags: [Cupons]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: integer }
                codigo: { type: string }
                tipo: { type: string }
                valor: { type: number }
                validoAte: { type: string, format: date-time }
      responses:
        '200':
          description: Cupom atualizado
    delete:
      summary: Deletar cupom
      tags: [Cupons]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [codigo]
              properties:
                codigo: { type: string }
      responses:
        '200':
          description: Cupom deletado
  /api/cupons/{codigo}/validar:
    post:
      summary: Validar cupom
      tags: [Cupons]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: codigo
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cupom válido
        '400':
          description: Cupom inválido, expirado ou sem usos

  # ====== NOTIFICAÇÕES ======
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginCredenciais:
      type: object
      required: [login, senha]
      properties:
        login: { type: string }
        senha: { type: string }
    LoginPin:
      type: object
      required: [pin]
      properties:
        pin: { type: string, maxLength: 4 }
...
    CriarCliente:
      type: object
      required: [nome]
      properties:
        nome: { type: string }
        cpf: { type: string }
        email: { type: string, format: email }
        telefone: { type: string }
        endereco: { type: string }
        obs: { type: string }
    AtualizarCliente:
      type: object
      required: [id]
      properties:
        id: { type: integer }
        nome: { type: string }
        cpf: { type: string }
        email: { type: string }
        telefone: { type: string }
        ativo: { type: boolean }